Complete the 23's exercises

+++ exercises/23_conversions/using_as.rs
@@ -10,24 +10,22 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
-
 fn average(values: &[f64]) -> f64 {
-    let total = values.iter().sum::<f64>();
-    total / values.len()
+  let total = values.iter().sum::<f64>();
+  total / values.len() as f64
 }
 
 fn main() {
-    let values = [3.5, 0.3, 13.0, 11.7];
-    println!("{}", average(&values));
+  let values = [3.5, 0.3, 13.0, 11.7];
+  println!("{}", average(&values));
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+  use super::*;
 
-    #[test]
-    fn returns_proper_type_and_value() {
-        assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
-    }
+  #[test]
+  fn returns_proper_type_and_value() {
+    assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
+  }
 }
